Fix 500 errors from upstream files not existing
Previously, trying to download a original or modified file based on an
upstream file that does not exist would result in a 500 error. Now we
check for that case and, if the file does not exist, we return a 404
error instead. Unit tests have been added to ensure this behaviour does
not regress.

Testing done:
Ran unit tests.

Testing Done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8591/