Allow search to be disabled
Previously, we always attempted to save the configuration form for the
currently selected search backend, even when we were disabling search.
This lead a 500 error and search not being able to be disabled (as we
were only checking the validity of the form when search was enabled).
Now we only save the sub-form configuration when search is enabled,
which lets search be disabled successfully.

Testing Done:
I was able to disable search with this patch applied. Without it I
could not.

Reviewed at https://reviews.reviewboard.org/r/8600/