Don't try to create an admin user if one already exists.
When setting up a site, it's possible that there will be existing data in the
database (for example, when migrating a server, or when an initial installation
attempt was cancelled). In this case, creating a superuser would fail due to a
duplicate username. This change makes it so we check first to verify if data
exists.

Testing done:
Created a test site.

Fixes bug 4487.

Reviewed at https://reviews.reviewboard.org/r/8618/