Fix a couple bugs caused by passing in unicode headers to urlopen.
`urlopen` is happy to accept a unicode object for headers, but it then tries to
use the ascii codec to encode them. If the text includes something that can't
be encoded, it would fail. This change fixes up a couple places where this was
causing problems for people.

Testing done:
Ran unit tests.

Fixes bugs 4069, 4472.

Reviewed at https://reviews.reviewboard.org/r/8619/