Add template hooks to email (html and text)
The two new hooks (one for html and one for text) can be used to add similar
information as what can be added to the review request pages using the
review-summary-header-pre/post template hooks.

The template hook requires the Django request object so that is now passed in
the review_published signal.

Testing Done:
Tested using an extension with template hooks for the new hooks. The hooks succesfully added information from the review extra_data field to the mail (both text and html).

Reviewed at https://reviews.reviewboard.org/r/8587/