Viewing a reply to a comment on the diff page now properly scrolls down to that reply
When clicking on the "View" button on a comment that is a reply to another
top-level comment, the browser redirected, but it didn't scroll down to that
specific reply you were trying to review.

I added an anchor tag in review_reply.html such that when clicking on the view
button, it now properly scrolls down to that comment.

Testing Done:
I reloaded the page on the dev-server (using MacOS Sierra, Chrome), clicked the
view button, and it redirected to the diff page and properly scrolled down to
the corresponding comment.

Bugs closed: 4085

Reviewed at https://reviews.reviewboard.org/r/8640/