Validate the bug tracker URLs for Jira and VersionOne.
Both Jira and VersionOne were allowing for URLs containing a "%s", which
were incorrect. The URLs were supposed to be the URLs to the service,
rather than a specific ticket. This ultimately led to crashes when
trying to navigate to the bug.

This change adds the missing validators for these fields.

Testing Done:
Tested that I could no longer specify bug trackers containing "%s", which
was otherwise causing crashes.

Reviewed at https://reviews.reviewboard.org/r/8664/