Fix up test code that creates a HostingServiceAccount.
The code that creates a new `HostingServiceAccount` for use in unit tests was
creating the object but not saving it, which would cause problems if we then
tried to create a relation to it when running within Django 1.8's atomic test
cases.

Testing done:
Ran unit tests on Django 1.6 and 1.8.

Reviewed at https://reviews.reviewboard.org/r/8700/