Fix up a couple unit tests that relied on database ordering.
There were a couple unit tests that checked the order of comments returned in a
query after saving. This test was somewhat fragile to the way that the database
backend would decide to return them, and had started failing for me. Since we
care specifically about the order of the timestamps, I've fixed up the tests to
provide `order_by('timestamp')` and the tests pass as expected.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8691/