Instantiate a Template instance when rendering custom webhook content.
We use Django's template language for allowing people to render custom webhook
content, but we deliberately limit the tags and filters that people are allowed
to use. Previously we just instantiated the `Parser` and told the resulting
nodelist to render, but in Django 1.8 there's some code that relies on having a
`Template` object for state. This change creates a `Template` and then replaces
its nodelist with our parsed version.

Testing done:
Ran unit tests with Django 1.6 and 1.8.

Reviewed at https://reviews.reviewboard.org/r/8702/