Add db_table and verbose_name for all models.
In preparation for making use of Django 1.8's AppConfig for all our apps (and
in doing so, rename them to be `reviewboard_*` instead of `*`), we need to
explicitly define the `db_table` for existing models so as to retain backward
compatibility.

While I was going through here, I also defined `verbose_name` and
`verbose_name_plural` for every model, setting them to be localized strings in
title case. We had a mix of things in here, often only defining
`verbose_name_plural` when we had something that needed to pluralize with "ies"
instead of "s" (and not localized). A recent change added these for the
`reviews` models, but these were all in lower case.

Testing done:
- Ran unit tests.
- Started the devserver and looked through the db admin to see the new casing
  on localized strings.

Reviewed at https://reviews.reviewboard.org/r/8704/