Use the Djblets MySQL backend in Review Board.
Djblets just gained a new and improved MySQL backend. This change
updates Review Board to make use of it when the standard Django MySQL
backend is configured.

settings.py now scans the database backends that are configured, looking
for the Django MySQL backend. If found, the backend is swapped out for
the Djblets one, requiring no change in the server's configuration.

Testing Done:
Tested with a standard Django MySQL backend in settings_local.py. Saw
that the Djblets backend was activated.

Bugs closed: 4007

Reviewed at https://reviews.reviewboard.org/r/8641/