Fix TestCase.create_webhook to set events as a list.
django-multiselectfield will do some comma-splitting when it loads a value from
the db, but if you set the value in the field to a string and immediately read
it back without committing to the db, it will return a string rather than a
list. In this case, webhooks that were created by a testcase running under
Django 1.8 (which dumps each test case into an atomic block) were failing
because the value read back was `'*'` instead of `['*']`.

Testing done:
Ran unit tests on Django 1.6 and 1.8.

Reviewed at https://reviews.reviewboard.org/r/8712/