Show the settings field when editing user profiles
When viewing user profiles inline in the database view, the `settings`
field was being omitted. This patch corrects that behaviour, adding it
under the set of state fields which are hidden by default.

Testing Done:
Viewed a user's profile on the user db page and saw the settings
field.

Reviewed at https://reviews.reviewboard.org/r/8715/