Fix display of "popup" pages in the admin site.
The admin site has a feature where it can show a pop-up window for some
actions, mostly selecting or creating related objects. We had some logic in the
`admin_actions` tag that was supposed to hide the sidebar in these windows, but
that logic was broken (since it was ORing two mutually-exclusive tests). I
fixed it to never even call `admin_actions` for pop-up windows, but that
exposed some issues with the CSS.

This change fixes all of that. The pop-up windows still show the header, which
I'm undecided on.

Testing done:
Opened a pop-up from the admin site and saw that it no longer showed the
sidebar.

Reviewed at https://reviews.reviewboard.org/r/8717/