Remove some (unused) relations from test_users.json
The test fixtures for users contained a couple items that had relations to
other things that didn't exist (a LocalSiteProfile and a ReviewRequestVisit).
In Django 1.6, loading a fixture with unresolved foreign keys works just fine,
but in 1.8, it throws an error. Because we don't actually use these particular
items in any tests, we can just get rid of them.

Testing done:
Ran unit tests on Django 1.6 and 1.8.

Reviewed at https://reviews.reviewboard.org/r/8718/