Convert URL patterns to lists.
Django has deprecated the use of the old `patterns()` method, since they now
recommend that all URLs pass in the view directly rather than do string magic
to get the module path. This change moves us over to using plain lists of
`url()` entries, rather than a bunch of calls to `patterns()`.

While I was doing this, I cleaned things up a bit with the regexes (such as
converting `[0-9]` to `\d` and `[a-zA-Z0-9_]` to `\w`), and changed it so we
only use raw strings where necessary.

Testing done:
Ran unit tests on Django 1.6 and 1.8.