Limit results from search resource.
On very large systems, calling the search resource with an empty `?q=` would
result in the resource returning every accessible user, group, and review
request on the system. This isn't a problem with the UI, since the javascript
only makes requests to this resource with non-empty queries, but it is a
possible way to DoS a server.

This change adds a `max-results` parameter, which defaults to 25 and maxes out
at 200. This mimics the way `max-results` works for list resources, with the
exception that there's no pagination (i.e. only the "first page" of results can
be returned). This change also adds documentation for the request fields that
GET on the search resource accepts.

Testing Done:
Ran unit tests.

Bugs closed: 4490

Reviewed at https://reviews.reviewboard.org/r/8581/