Allow inactive users to be fetched from the API
Previously attempting to retrieve a user from the API who had been set
to inactive would result in a 404 Not Found, which was an issue for any
review requests or reviews they had created (as the link included would
not work). Now we return a user from the `/api/users/<username>/` API
regardless of whether or not they are active. However, we still filter
out users from the `/api/users/` API unless `include-inactive=1` is
provided as a GET parameter to the request.

Testing done:
Ran unit tests.

Reviewed at https://reviews.reviewboard.org/r/8667/