Fix a major regression in comment ordering.
A recent change going into 2.0.26 and 2.5.8 broke comment ordering,
causing them to be ordered by ID and not timestamp. This could greatly
alter reading and composing a review, along with API results. The cause
was the subclasses of BaseComment introducing a Meta that overrode the
one from the parent, instead of subclassing it.

This fixes that by subclassing the parent's meta, inheriting the
ordering. It also removes the app_labels, which were already provided by
the parent.

Testing Done:
Unit tests pass.

No longer saw the comment ordering problem in tests.

Reviewed at https://reviews.reviewboard.org/r/8739/