Do not render the avatar service form when there are no services
We now take into account the number of enabled services when calculating
the visibility of the avatar settings form. In addition, due to recent
changes to Djblets (see /r/8726 and /r/8728), we no longer have to add or
remove the avatar services form to the configuration page (due to
`ConfigPageForm.is_visible()` evlauating to `False` causing the form to
not render). This allows the logic of the avatar services admin form to
be simplified.

Testing Done:
Viewed account page with avatars enabled but no enabled services. The
page did not raise an exception and the form was not visible.

Reviewed at https://reviews.reviewboard.org/r/8729/