Fix a regression with using API Tokens to log in.
Review Board 2.5.8 introduced the ability to log into the API using an
e-mail address instead of a username, but this broke the ability to log
in using API tokens. An assumption was made that the credentials for an
auth request had a username, but this isn't true for API tokens.

This change wires off the new logic if a username is not provided.

Testing Done:
Successfully logged in using API tokens and using e-mail addresses.

Bugs closed: 4521

Reviewed at https://reviews.reviewboard.org/r/8743/