Add old status/public states to the review request reopened signal.
This fleshes out the review request reopened signals to include
information on what the status and public states were prior to
reopening. Consumers can use this to more intelligently determine
whether to take action on the reopen. For instance, it could decide not
to send a message to a service for a review request that was previously
unpublished and is now being reopened, or it could send a different
message when reopened from a discarded state vs. a submitted state.

Unit tests were added to test reopen functionality and resulting states,
including the new signal changes.

Testing Done:
Unit tests pass.

Reviewed at https://reviews.reviewboard.org/r/8724/