Bump the minimum version of pytz to 2015.1.
Someone on an older system hit a problem in Review Board 2.5.8 involving
a failing import from an old (2010) version of pytz. Back then, the
pytz.exceptions module didn't exist.

This change bumps up our minimum version to 2015.1. Newer versions are
ideal, but this version is known to work (as we use it on one of our
servers), and should make it easier for packagers. Users upgrading from
an older version will get the latest version of this.

Testing Done:
Installed a Review Board package and got the latest version.

Verified that 2015.1 has the module we require.

Bugs closed: 4519

Reviewed at https://reviews.reviewboard.org/r/8751/