Add release notes for Review Board 2.0.27.
This is a small release that fixes a comment ordering regression found
in 2.0.26, along with clipboard bugs in the diff viewer, My Account form
visibility bugs for extensions, and crashes with certain logging calls
on older versions of Python.

Testing Done:
Built and checked the docs.

Reviewed at https://reviews.reviewboard.org/r/8744/