Move API resource version handling into the API doc generator.
We were keeping information on resource versioning (when it was added,
deprecated, removed) in two places: In the resource and in the doc file.
While we haven't had any inconsistencies, it doesn't make sense to have
this in two places when we can easily generate what we want in the docs
generator based on the metadata on the resources.

This adds support in the API doc generator to output the versioning
information, and removes the manual information from each of the files,
simplifying them.

It also introduces "item_resource_" and "list_resource_" prefixes for
the "added_in", etc. versioning fields, for the cases (one case,
currently) where lists and items weren't added in the same version.

Testing Done:
Built the docs and checked that all generated docs with resources
referencing version information had the information listed under
the resource name.

Verified that each resource and each doc file had matching version
information before removing from the files.

Reviewed at https://reviews.reviewboard.org/r/8770/