Fix conflicts that could occur with move flag anchors.
The move flags had anchors set to jump to the other side of a move when
clicked. These were using pure line numbers, which could cause conflicts
when referencing a line that another move block in the same file started
at. For example, a move flag in the left-hand side of the diff viewer
that referenced line 100 in the right-hand side could jump to the wrong
flag if the left-hand side also had a move flag starting at its own line
100.

Fixing this is a simple matter of namespacing the anchors so that it
specifically references the correct side of the diff.

Testing Done:
Tested that comment flags were working correctly and not conflicting with
each other.

Reviewed at https://reviews.reviewboard.org/r/8796/