Fix duplicating the "whitespace changes only" banner when expanding lines.
Files that just have whitespace-only changes have a yellow banner saying
"This file contains only whitespace changes." This was being
unconditionally generated in the template for the first chunk in a file,
which meant that if you expanded the lines for the first chunk, you'd
get another copy of the banner, and same with collapsing.

This change adds a check in the template to prevent this banner from
appearing in "standalone" mode (where chunks of a diff are being
returned without the rest of the file).

It also adds a bottom border to this, so that when the chunk is
expanded, it has a nice clean separation from any equals lines.

Testing Done:
Tested with a file containing only whitespace changes. Expanding and
collapsing the first chunk didn't result in duplicate banners.

Bugs closed: 4422

Reviewed at https://reviews.reviewboard.org/r/8799/