Fix a regression in the Remote Resource API from recent changes.
The recent API documentation changes make a modification to
RemoteRepositoryResource.get_list(), which changed it to not augment the
parent resource in order to avoid inheriting some of the API metadata
for the request. However, it didn't fix the method body to call the
parent method, breaking GET requests on this resource.

This is a trivial change that adds a super call.