Avoid error message on "New Review Request page" for Mercurial < 3.9
If the repository hosting uses a Mercurial 3.7 it is possible to get branches
(/json-branches/) with the JSON API of hgweb. But it leads to an error if
reviewboard tries to fetch "/json-log/?rev=" as it is not implemented.

Also avoid an error if Mercurial is really old and does not support to get
branches in JSON style with "/json-branches/".

Testing Done:
Opened "New Review Request page" and check if branch and commit list works
properly and does not contain errors.

Tested versions: 2.6.2, 3.0, 3.7.3, 4.1

Bugs closed: 4524

Reviewed at https://reviews.reviewboard.org/r/8778/