Make our python-dateutil dependency less strict.
A long time ago, we capped out python-dateutil dependency to 1.5, in
order to retain support for Python 2.x. This was because dateutil 2.0
only supported Python 3. However, dateutil 2.1+ supports Python 2.6 and
higher.

We're now bumping up the dependency to support 1.5 and higher. We can't
explicitly forbid 2.0 support, but if installing via a Python package,
people will either end up with a modern version or will stay on 1.5.
Distro packagers can ship an appropriate version (and likely will have a
non-2.0 version anyway).

Testing Done:
Deployed a new build. Saw that it worked fine with the newer dateutil 2.6
that was on the server, and didn't trigger dependency errors.

Reviewed at https://reviews.reviewboard.org/r/8838/