Fix a JavaScript issue involving "use strict" and class declarations.
A recent deployment of Review Board to a new server unveiled a breakage
with our JavaScript for 3.0. The new JavaScript is being built with
"use strict" mode, and this has so far uncovered one regression, which
were some banners that were defined but not explicitly declared with a
"var" statement. While we've had this for a long time, it never came up
due to "use strict" not formerly being used.

This is something we may want to investigate shutting off temporarily
for the 3.0 beta 1 release, but continue to test internally and on our
server.

This change adds the missing variable declarations, fixing the diff
viewer.

Testing Done:
Uploaded a fixed build to reviews.reviewboard.org. I was able to view
diffs again.

Reviewed at https://reviews.reviewboard.org/r/8837/