Fix cache issues with status updates.
If a status update is updated (for example, going from starting -> done)
without an associated review, timestamps wouldn't be updated such that the
review request ETag would change. To fix that, I've added `auto_now` to the
timestamp field so that it gets updated whenever the status update is changed,
and fixed the ETag calculation to look at all the status update timestamps.

Testing done:
Posted a review and had Review Bot start. Saw that the initial status update
was marked as "starting", and then reloaded. Whereas before it would still show
as starting (without a shift+reload), now it showed that the tool ran
sucessfully.

Reviewed at https://reviews.reviewboard.org/r/8856/