Fix resource paths for user file attachments.
When I fixed the names in the user file attachments resource to make it
available within the API templates, I neglected to actually test the javascript
in the Review Board UI, which was still using the old paths and names. This
change fixes it.

Testing done:
Drag-and-dropped an image into a review header.

Reviewed at https://reviews.reviewboard.org/r/8860/