Support managing OAuth2 applications in the db
ReviewBoard now depends on `django-oauth-toolkit` to provide support as
an OAuth2 identity provider for OAuth2 client applications. This first
patch allows applications to be managed in the database by
administrators. We use our own custom application model to allow
applications to be limited to `LocalSite`s, as well as for attaching
extra data fields to them.

This patch also provides help text for the application form fields, most
of which were missing them.

Testing Done:
I was able to create an application in the database successfully.

Reviewed at https://reviews.reviewboard.org/r/8630/